improve protected-route naming & file structure
