+++ src/components/unauthorized-access-only/unauthorized-access-only.jsx
@@ -1,0 +1,25 @@
+// libraries
+import PropTypes from "prop-types";
+
+// components
+import AuthorizedAccessOnly from "../authorized-access-only/authorized-access-only";
+
+
+
+function UnauthorizedAccessOnly({ element }) {
+  return (
+    <AuthorizedAccessOnly
+      element={element} 
+      onlyUnauthorized={true} 
+    />
+  );
+};
+
+
+UnauthorizedAccessOnly.propTypes = {
+  element: PropTypes.element.isRequired
+};
+
+
+
+export default UnauthorizedAccessOnly;
